+++ packages/blend/lib/components/PopoverV2/popoverV2.types.ts
@@ -62,7 +62,7 @@ export type PopoverV2Props = {
     // zIndex?: number
     size?: PopoverV2Size
     onClose?: () => void
-    // shadow?: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | 'full'
+    shadow?: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | 'full'
     useDrawerOnMobile?: boolean
     avoidCollisions?: boolean
     skeleton?: PopoverV2SkeletonProps